requirements.txt
cryptography>=39.0.0
pymysql
pycryptodome
python-dotenv
click
